* commit 'cbf339f6e4bf45a7d41cd0327d2dd7975875b8aa':
  AArch64: Use long for pointers in services

+++ services/java/com/android/server/input/InputWindowHandle.java
@@ -28,7 +28,7 @@ public final class InputWindowHandle {
     // Pointer to the native input window handle.
     // This field is lazily initialized via JNI.
     @SuppressWarnings("unused")
-    private int ptr;
+    private long ptr;
 
     // The input application handle.
     public final InputApplicationHandle inputApplicationHandle;
